requirements.txt
flask
langchain
langchain-google-genai
langchain-pinecone
google-generativeai
python-dotenv
pinecone-client
gunicorn 